+++ learning/katas/java/IO/Built-in IOs/Built-in IOs/test/org/apache/beam/learning/katas/io/builtinios/TaskTest.java
@@ -16,19 +16,7 @@
  * limitations under the License.
  */
 
-import org.apache.beam.sdk.Pipeline;
-import org.apache.beam.sdk.options.PipelineOptions;
-import org.apache.beam.sdk.options.PipelineOptionsFactory;
+package org.apache.beam.learning.katas.io.builtinios;
 
-public class Task {
-
-  public static void main(String[] args) {
-    PipelineOptions options = PipelineOptionsFactory.fromArgs(args).create();
-    Pipeline pipeline = Pipeline.create(options);
-
-
-
-    pipeline.run();
-  }
-
+public class TaskTest {
 }
